+++ src/characters/Sardine.ts
@@ -7,7 +7,8 @@ class Sardine extends Character {
   protected readonly _isBenthic: boolean = false;
   protected readonly _imageFilename: string = "sardine.png";
   protected readonly _foodValue: number = 0.6;
-  protected readonly _damage: number;
+  protected readonly _damage: number = 0;
+  protected readonly _speed: number = 1.3; 
 }
 
 export default Sardine;