+++ src/components/Footer.tsx
@@ -3,7 +3,7 @@ import { version } from "../../package.json";
 
 function Footer() {
   return (
-    <div className="fixed bottom-1 right-1">
+    <div className="fixed bottom-1 right-1 flex gap-5">
       <a
         href="https://github.com/danieldesira/TurtleEscape/blob/master/changelog.md"
         target="_blank"