+++ src/characters/NeptuneGrass.ts
@@ -10,6 +10,7 @@ class NeptuneGrass extends Character {
   protected readonly _foodValue: number = 5;
   protected readonly _damage: number = 0;
   protected readonly _speed: number = 0;
+  protected readonly _stomachImpact: number = 10;
 }
 
 export default NeptuneGrass;