+++ src/characters/PlasticBag.ts
@@ -9,7 +9,8 @@ class PlasticBag extends Character {
   protected readonly _imageFilename: string = "plastic-bag.png";
   protected readonly _foodValue: number = 0;
   protected readonly _damage: number = 8;
-  protected readonly _speed: number = 0.1; 
+  protected readonly _speed: number = 0.1;
+  protected readonly _stomachImpact: number = 20;
 }
 
 export default PlasticBag;