+++ src/characters/Sardine.ts
@@ -9,7 +9,8 @@ class Sardine extends Character {
   protected readonly _imageFilename: string = "sardine.png";
   protected readonly _foodValue: number = 0.6;
   protected readonly _damage: number = 0;
-  protected readonly _speed: number = 1.3; 
+  protected readonly _speed: number = 1.3;
+  protected readonly _stomachImpact: number = 1;
 }
 
 export default Sardine;