+++ src/features/turtleMonitor/TurtleState.ts
@@ -3,6 +3,7 @@ interface TurtleState {
     oxygen: { value: number };
     life: { value: number };
     food: { value: number };
+    stomachCapacity: { value: number };
   };
 }